requirements.txt
asgiref==3.8.1
Django==4.2.11
pillow==10.3.0
psycopg2-binary==2.9.9
pypng==0.20220715.0
python-barcode==0.15.1
python-dotenv==1.0.1
qrcode==7.4.2
sqlparse==0.5.0
typing_extensions==4.11.0
